removed json

+++ action_run_all_to_update_messy.py
@@ -5,11 +5,34 @@ import time
 
 
 def main(**kwargs):
+
+    #mode = 'full'
+    mode = 'fast'
+
+    repo_names = {}
+    current_names = []
+    current_names.append('oomlout_oomp_version_1_only_yaml')
+    current_names.append('oomlout_oomp_current_version_messy')    
+    repo_names['full'] = current_names
+
+    current_names = []
+    current_names.append('oomlout_oomp_version_1_test_fast_only_yaml')
+    current_names.append('oomlout_oomp_version_1_test_fast_messy')
+    repo_names['fast'] = current_names
+
+    all_repos = []
+    for key in repo_names:
+        all_repos += repo_names[key]
+    repo_names['all'] = all_repos
+
+
+
+
     # Store the start time 
     start_time = time.time()
 
-    # Run the creation process
-    if True:
+    # Run the creation process skip of going fast
+    if True and mode == 'full':
         pass
         os.chdir('C:\\gh\\oomlout_oomp_part_generation_version_1')
         subprocess.run(['python', 'action_setup_full.py'])
@@ -21,9 +44,12 @@ def main(**kwargs):
     #delete the old oomp
     if True:
         print('Deleting old oomp')
+        #directory_to_delete = []
+        #directory_to_delete.append('z:\\oomlout_oomp_current_version_messy')
+        #directory_to_delete.append('z:\\oomlout_oomp_version_1_only_yaml')
         directory_to_delete = []
-        directory_to_delete.append('z:\\oomlout_oomp_current_version_messy')
-        directory_to_delete.append('z:\\oomlout_oomp_version_1_only_yaml')
+        for repo_name in all_repos:
+            directory_to_delete.append(f'z:\\{repo_name}')  
         delete_directory(directory_to_delete)
 
         
@@ -32,7 +58,11 @@ def main(**kwargs):
     #run only_yaml
     if True:
         #add  Z:\oomlout_oomp_current_version_messy to python path
-        repo_name = 'oomlout_oomp_version_1_only_yaml'
+        if mode == 'full':
+            repo_name = 'oomlout_oomp_version_1_only_yaml'
+        elif mode == 'fast':
+            repo_name = 'oomlout_oomp_version_1_test_fast_only_yaml'
+
         repo_url = f"https://github.com/oomlout/{repo_name}"
         directory_current = f'Z:\\{repo_name}'
         
@@ -66,7 +96,10 @@ def main(**kwargs):
     #run messy
     if True:
         #add  Z:\oomlout_oomp_current_version_messy to python path
-        repo_name = 'oomlout_oomp_version_1_messy'
+        if mode == 'full':
+            repo_name = 'oomlout_oomp_version_1_messy'
+        elif mode == 'fast':
+            repo_name = 'oomlout_oomp_version_1_test_fast_messy'
         repo_url = f"https://github.com/oomlout/{repo_name}"
         directory_current = f'Z:\\{repo_name}'
         
@@ -114,18 +147,40 @@ def delete_directory(directory_to_delete):
     if not isinstance(directory_to_delete, list):
         directory_to_delete = [directory_to_delete]
     for directory in directory_to_delete:
-        try:
-            print(f'Deleting {directory}')
-            # Delete directory and all files and subdirectories within using os.system in Windows
-            result = os.system(f'rmdir /s /q {directory}')
-            if result != 0:
-                raise Exception(f'Failed to delete {directory}')
-        except Exception as e:
-            print(e)
+        #if directory exists
+        if not os.path.exists(directory):
+            print(f'{directory} does not exist')
+        else:
             try:
-                shutil.rmtree(directory)
-            except Exception as shutil_error:
-                print(f'Failed to delete {directory} using shutil: {shutil_error}')
+                print(f'Deleting {directory}')
+                # Delete directory and all files and subdirectories within using os.system in Windows
+                result = os.system(f'rmdir /s /q {directory}')
+                if result != 0:
+                    raise Exception(f'Failed to delete {directory}')
+            except Exception as e:
+                print(e)
+                try:
+                    shutil.rmtree(directory)
+                except Exception as shutil_error:
+                    print(f'Failed to delete {directory} using shutil: {shutil_error}')
+
+            # remvove empty directory
+            if not os.path.exists(directory):
+                    print(f'{directory} does not exist')
+            else:
+                print(f'Waiting 20 seconds before trying to remove empty directory {directory}')
+                time.sleep(20)
+                try:
+                    os.rmdir(directory)
+                except Exception as e:
+                    #try os.system
+                    command = f'rmdir /s /q {directory}'
+                    print(f'Running command: {command}')
+                    result = os.system(command)
+                    if result != 0:
+                        print(f'Failed to remove empty directory {directory}')
+                        pass#time.sleep(30)
+            
 
 
 if __name__ == '__main__':
